Online Medicine Booking
The main objective of this system is to take the whole medical store online so that it is reachable to customers 24/7. It also aims at going towards cashless transactions. It will impart a wider visibility to the customers. Thus boosting the business to higher levels. The software will be a web based system with a very user friendly interface which indeed will make the whole management process easy to manage and operate with zero redundancies. Overall online medical booking store will become an efficient, highly responsive and an extremely accurate system.